extract parser into its own file
